Use M3U8LineReader when reading master playlist
The existing logic breaks when `\r` is present in the playlist. There was special handling for the URI line, but in certain cases (example: audio stream with URI at the end) it broke other lines too.

M3U8LineReader is already used when reading media playlists and solves the issue in a clean way.